bot.py
import os
import pandas as pd
import datetime
import json
from telegram import Update, ReplyKeyboardMarkup
from telegram.ext import Application, CommandHandler, MessageHandler, filters, ContextTypes

class AntiProcrastinatorBot:
    def __init__(self):
        self.token = os.environ.get("TELEGRAM_TOKEN", "8324093095:AAGMioQcN_A0fDi9o85c2y9N0EzhjAciCJA")
        
        # Для облака используем переменные окружения или встроенную БД
        self.setup_cloud_storage()
        
        print("🚀 Бот запущен в облаке 24/7!")
    
    def setup_cloud_storage(self):
        """Настройка хранилища для облака"""
        try:
            # Попробуем использовать SQLite для простоты
            import sqlite3
            self.conn = sqlite3.connect('bot_data.db', check_same_thread=False)
            self.create_tables()
            self.storage_type = "sqlite"
        except:
            # Файловое хранилище как запасной вариант
            self.storage_type = "json"
    
    def create_tables(self):
        cursor = self.conn.cursor()
        cursor.execute('''
            CREATE TABLE IF NOT EXISTS users (
                user_id INTEGER PRIMARY KEY,
                username TEXT,
                registration_date TEXT,
                last_active TEXT
            )
        ''')
        cursor.execute('''
            CREATE TABLE IF NOT EXISTS tasks (
                id INTEGER PRIMARY KEY AUTOINCREMENT,
                user_id INTEGER,
                task_text TEXT,
                created_date TEXT,
                completed BOOLEAN DEFAULT FALSE
            )
        ''')
        self.conn.commit()

    # ... остальные методы адаптированные для SQLite ...

# requirements.txt для облака: