requirements.txt.txt
pandas==2.0.3
python-telegram-bot==20.7